+++ pytests/test_testing.py
@@ -20,7 +20,7 @@ def test_ffwd_iter():
 
 def test_testing_source():
     inp = TestingSource(range(3))
-    part = inp.build_part("iterable", None)
+    part = inp.build_part("test", "iterable", None)
     assert part.next_batch() == [0]
     assert part.next_batch() == [1]
     assert part.next_batch() == [2]
@@ -31,7 +31,7 @@ def test_testing_source():
 
 def test_testing_source_resume_state():
     inp = TestingSource(range(3))
-    part = inp.build_part("iterable", None)
+    part = inp.build_part("test", "iterable", None)
     assert part.next_batch() == [0]
     assert part.next_batch() == [1]
     resume_state = part.snapshot()
@@ -40,7 +40,7 @@ def test_testing_source_resume_state():
     part.close()
 
     inp = TestingSource(range(3))
-    part = inp.build_part("iterable", resume_state)
+    part = inp.build_part("test", "iterable", resume_state)
     assert part.snapshot() == resume_state
     assert part.next_batch() == [2]
     with raises(StopIteration):
@@ -50,7 +50,7 @@ def test_testing_source_resume_state():
 
 def test_testing_source_batch_size():
     inp = TestingSource(range(5), batch_size=2)
-    part = inp.build_part("iterable", None)
+    part = inp.build_part("test", "iterable", None)
     assert part.next_batch() == [0, 1]
     assert part.next_batch() == [2, 3]
     assert part.next_batch() == [4]
@@ -59,7 +59,7 @@ def test_testing_source_batch_size():
 
 def test_testing_source_eof():
     inp = TestingSource([0, 1, 2, TestingSource.EOF(), 3, 4], batch_size=2)
-    part = inp.build_part("iterable", None)
+    part = inp.build_part("test", "iterable", None)
     assert part.next_batch() == [0, 1]
     assert part.next_batch() == [2]
     with raises(StopIteration):
@@ -67,7 +67,7 @@ def test_testing_source_eof():
     part.close()
 
     resume_state = part.snapshot()
-    part = inp.build_part("iterable", resume_state)
+    part = inp.build_part("test", "iterable", resume_state)
     assert part.next_batch() == [3, 4]
     with raises(StopIteration):
         part.next_batch()
@@ -92,14 +92,14 @@ def test_testing_source_eof_run(recovery_config):
 
 def test_testing_source_abort():
     inp = TestingSource([0, 1, 2, TestingSource.ABORT(), 3, 4], batch_size=2)
-    part = inp.build_part("iterable", None)
+    part = inp.build_part("test", "iterable", None)
     assert part.next_batch() == [0, 1]
     resume_state = part.snapshot()
     assert part.next_batch() == [2]
     with raises(AbortExecution):
         part.next_batch()
 
-    part = inp.build_part("iterable", resume_state)
+    part = inp.build_part("test", "iterable", resume_state)
     assert part.next_batch() == [2, 3]
     assert part.next_batch() == [4]
     with raises(StopIteration):