+++ src/inputs.rs
@@ -216,11 +216,16 @@ impl FixedPartitionedSource {
     fn build_part(
         &self,
         py: Python,
+        step_id: &StepId,
         for_part: &StateKey,
         resume_state: Option<TdPyAny>,
     ) -> PyResult<StatefulPartition> {
         self.0
-            .call_method1(py, "build_part", (for_part.clone(), resume_state))?
+            .call_method1(
+                py,
+                "build_part",
+                (step_id.clone(), for_part.clone(), resume_state),
+            )?
             .extract(py)
     }
 
@@ -281,6 +286,10 @@ impl FixedPartitionedSource {
             .f64_histogram("inp_part_next_batch_duration_seconds")
             .with_description("`next_batch` duration in seconds")
             .init();
+        let batch_size_histogram = meter
+            .u64_histogram("inp_part_next_batch_size")
+            .with_description("`next_batch` batch size")
+            .init();
         let snapshot_histogram = meter
             .f64_histogram("snapshot_duration_seconds")
             .with_description("`snapshot` duration in seconds")
@@ -331,6 +340,7 @@ impl FixedPartitionedSource {
                                     tracing::info!("Resuming {part_key:?} at epoch {emit_epoch} with state {state:?}");
                                     let part = self.build_part(
                                         py,
+                                        &step_id,
                                         &part_key,
                                         Some(state)
                                     ).reraise_with(|| format!("error calling `FixedPartitionSource.build_part` in step {step_id} for partition {part_key}"))?;
@@ -392,7 +402,7 @@ impl FixedPartitionedSource {
                                 for part_key in &primary_parts {
                                     if !parts.contains_key(part_key) {
                                         tracing::info!("Init-ing {part_key:?} at epoch {epoch:?}");
-                                        let part = self.build_part(py, part_key, None)
+                                        let part = self.build_part(py, &step_id, part_key, None)
                                             .reraise_with(|| format!("error calling `FixedPartitionSource.build_part` in step {step_id} for partition {part_key}"))?;
                                         let next_awake = part.next_awake(py)
                                             .reraise_with(|| format!("error calling `StatefulSourcePartition.next_awake` in step {step_id} for partition {part_key}"))?;
@@ -450,6 +460,7 @@ impl FixedPartitionedSource {
                                         match batch_res {
                                             BatchResult::Batch(batch) => {
                                                 let batch_len = batch.len();
+                                                batch_size_histogram.record(batch_len as u64, &labels);
 
                                                 let mut downstream_session = downstream_handle.session(&part_state.downstream_cap);
                                                 item_out_count.add(batch_len as u64, &labels);
@@ -664,11 +675,12 @@ impl DynamicSource {
     fn build(
         &self,
         py: Python,
+        step_id: &StepId,
         index: WorkerIndex,
         count: WorkerCount,
     ) -> PyResult<StatelessPartition> {
         self.0
-            .call_method1(py, "build", (index.0, count.0))?
+            .call_method1(py, "build", (step_id.0.clone(), index.0, count.0))?
             .extract(py)
     }
 
@@ -693,7 +705,7 @@ impl DynamicSource {
         let worker_index = scope.w_index();
         let worker_count = scope.w_count();
         let part = self
-            .build(py, worker_index, worker_count)
+            .build(py, &step_id, worker_index, worker_count)
             .reraise_with(|| format!("error calling `DynamicSource.build` in step {step_id}"))?;
 
         let op_name = format!("{step_id}.dynamic_input");