+++ src/lib.rs
@@ -31,6 +31,5 @@ fn mod_bytewax(py: Python, m: &PyModule) -> PyResult<()> {
     run::register(py, m)?;
     tracing::register(py, m)?;
     window::register(py, m)?;
-
     Ok(())
 }