+++ CHANGELOG.md
@@ -9,6 +9,14 @@ For help with updating to new Bytewax versions, please see the
 __Add any extra change notes here and we'll put them in the release
 notes on GitHub when we make a new release.__
 
+- *Breaking change* `FixedPartitionedSource.build_part`,
+  `DynamicSource.build`, `FixedPartitionedSink.build_part` and `DynamicSink.build`
+  now take an additional `step_id` argument. This argument can be used when
+  labeling custom Python metrics.
+
+- Custom Python metrics can now be collected using the `prometheus-client`
+  library.
+
 - *Breaking change* The schema registry interface has been removed.
   You can still use schema registries, but you need to instantiate
   the (de)serializers on your own. This allows for more flexibility.
@@ -18,7 +26,6 @@ notes on GitHub when we make a new release.__
 - Fixes bug where items would be incorrectly marked as late in sliding
   and tumbling windows in cases where the timestamps are very far from
   the `align_to` parameter of the windower.
-
 - Adds `stateful_flat_map` operator.
 
 - *Breaking change* Removes `builder` argument from `stateful_map`.